flask-app/requirements.txt
flask
google-cloud-storage
Pillow
